pyproject.toml
[project]
name = "climaplan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "guardrails-ai>=0.7.2",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.5",
    "mcp>=1.25.0",
    "mcp-weather-server>=0.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
